floor model update

+++ app/django/items/models.py
@@ -32,8 +32,8 @@ class UnitType(models.Model):
 class UnitFloorType(models.Model):  # 층별 타입
     project = models.ForeignKey('project.Project', on_delete=models.CASCADE, verbose_name='프로젝트', related_name='floors')
     sort = models.CharField('타입종류', max_length=1, choices=UnitType.SORT_CHOICES)
-    start_floor = models.PositiveIntegerField('시작 층')
-    end_floor = models.PositiveIntegerField('종료 층')
+    start_floor = models.SmallIntegerField('시작 층')
+    end_floor = models.SmallIntegerField('종료 층')
     extra_cond = models.CharField('방향/위치', max_length=20, blank=True,
                                   help_text='동일범위의 층범위를 방향/위치 등으로 구분해야 할 필요가 있는 경우 입력')
     alias_name = models.CharField('층별 범위 명칭', max_length=20)
